3-RSS Parallel Mechanism for Neck (#30)
* refactor double to float

* refactor for multiple maestro boards

* add RSS code

* only accept connections from those in the allow_list, turn off scanning

* adjust DOME controller settings (Gray Left)

* minimize traffic to Maestro, only send commands when a change is detected

* turn off debugging

* reduce jumpiness of joycon input, deadband and only hundreths resolution

* adjusting data type conversions

+++ main/chopper/drive/SingleDriveSabertooth.cpp
@@ -5,10 +5,12 @@ void SingleDriveSabertooth::SetRampingValue(int ramping) {
     SingleDrive::SetRampingValue(ramping);
 }
 
-void SingleDriveSabertooth::SetDeadband(double deadband) {
+void SingleDriveSabertooth::SetDeadband(float deadband)
+{
     SingleDrive::SetDeadband(deadband);
 }
 
-void SingleDriveSabertooth::SetExpiration(uint64_t expirationTime) {
+void SingleDriveSabertooth::SetExpiration(uint64_t expirationTime)
+{
     SingleDrive::SetExpiration(expirationTime);
 };
