+++ main/include/chopper/core/ButtonState.h
@@ -30,7 +30,7 @@ public:
     void updateState(bool button_down)
     {
 
-        unsigned long currentMillis = Timer::GetFPGATimestamp();
+        uint64_t currentMillis = Timer::GetFPGATimestamp();
         if (button_down)
         {
             if (!is_pressed)
@@ -57,44 +57,44 @@ public:
         }
     }
 
-    uint32_t lastPressTime() const
+    uint64_t lastPressTime() const
     {
         return last_press_time;
     }
 
-    uint32_t lastReleaseTime() const
+    uint64_t lastReleaseTime() const
     {
         return last_release_time;
     }
     
-    uint32_t pressedDuration() const
+    uint64_t pressedDuration() const
     {
-        unsigned long currentMillis = Timer::GetFPGATimestamp();
+        uint64_t currentMillis = Timer::GetFPGATimestamp();
         return ((state_ & PRESSED) ? (currentMillis - last_press_time) : 0);
     }
 
     bool isPressedFor(uint32_t holdThreshold = 1000) const
     {
-        unsigned long currentMillis = Timer::GetFPGATimestamp();
+        uint64_t currentMillis = Timer::GetFPGATimestamp();
         return pressedDuration() >= holdThreshold;
     }
 
-    uint32_t releasedDuration() const
+    uint64_t releasedDuration() const
     {
-        unsigned long currentMillis = Timer::GetFPGATimestamp();
+        uint64_t currentMillis = Timer::GetFPGATimestamp();
         return ((state_ & PRESSED) == 0 ? (currentMillis - last_release_time) : 0);
     }
 
     bool isReleasedFor(uint32_t releaseThreshold = 1000) const
     {
-        unsigned long currentMillis = Timer::GetFPGATimestamp();
+        uint64_t currentMillis = Timer::GetFPGATimestamp();
         return releasedDuration() >= releaseThreshold;
     }
 
 private:
     int double_click_threshold_ms;
-    uint32_t last_press_time;
-    uint32_t last_release_time;
+    uint64_t last_press_time;
+    uint64_t last_release_time;
     bool is_pressed;
     uint8_t state_;              // Bitmask for button state
 };
